Consolidate wrapper locations
#47
